:truck: Move dates and conditions components to dates-and-conditions folder.
Why? the sub-folder should match the component name.
